evas: add an API to know where inside an object content can be layout over.
The first step with this API is to use the information provided when setting
the border on an image to define the content area inside it. Improvement will
be to use more flexible stretch zone area to make it more customizable and
finally read the information from a 9patch file (.9.png).