csharp: Add helpers to get/set Values from Objects
The user can construct an `Eina.Value` from a plain C# `object`, using
reflection to get the correct type of object and construct the correct
underlying C value.

Also added the `Unwrap()` method to return a C# object representing the
wrapped value.

Currently, containers are not supported. It will be added in a following
commit.